Overall rating display and comments

+++ bookworm/Structs/Review.swift
@@ -12,12 +12,12 @@ import FirebaseFirestore
 struct Review: Decodable {
     var userId: String
     var comment: String?
-    var rating: Int?
+    var rating: Double?
     var date: Date
     init(aDoc: DocumentSnapshot) {
         self.userId = aDoc.get("user_id") as? String ?? "Unknown"
         self.comment = aDoc.get("comment") as? String
-        self.rating = aDoc.get("rating") as? Int
+        self.rating = aDoc.get("rating") as? Double
         let ts = aDoc.get("date") as? Timestamp
         self.date = ts!.dateValue()
     }
